Split extract.ts into per-source modules under extract/
The single ~350-line extractor was about to compound — every v0.1.2
extraction-breadth issue (#69 recursive workspaces, #70 generalised DB
discovery, #71 calls beyond HTTP, #72 Python, #73 infra) wants its own
phase or its own source-type module. Doing the split first keeps each
later PR honest about its surface area.

Layout under packages/core/src/extract/:

  shared.ts     — DiscoveredService, IGNORED_DIRS, cleanVersion, exists,
                  read helpers, makeEdgeId
  services.ts   — phase 1: discoverServices + addServiceNodes
  databases.ts  — phase 2: extractDatabaseFromConfig, attachIncompatibilities,
                  addDatabasesAndCompat orchestration
  configs.ts    — phase 3: walkConfigFiles + addConfigNodes
  calls.ts      — phase 4: tree-sitter URL-literal scan + addCallEdges
  index.ts      — extractFromDirectory orchestrator (27 lines)

extract.ts at the package root becomes a one-line re-export so every
existing import path keeps working with no churn.

Behaviour is unchanged: the demo extraction produces the same nodes,
edges, and provenance as before. All 102 core tests pass without
modification, and `node packages/core/dist/cli.cjs init ./demo` produces
the identical incompatibility output.

Refs #68
